Update tmp_dir loc for metascope_id

+++ R/metascope_id.R
@@ -383,7 +383,7 @@ metascope_id <- function(input_file, input_type = "csv.gz",
                          db_feature_table = NULL,
                          NCBI_key = NULL,
                          out_dir = dirname(input_file),
-                         tmp_dir = NULL,
+                         tmp_dir = dirname(input_file),
                          convEM = 1 / 10000, maxitsEM = 25,
                          update_bam = FALSE,
                          num_species_plot = NULL,
@@ -405,6 +405,8 @@ metascope_id <- function(input_file, input_type = "csv.gz",
     blast_fastas <- FALSE
     update_bam <- FALSE
   }
+  # Check that tmp_dir is specified
+  if (is.null(tmp_dir)) stop("Please supply a directory for 'tmp_dir' argument.")
   reads <- obtain_reads(input_file, input_type, aligner, blast_fastas, quiet)
   unmapped <- is.na(reads[[1]]$rname)
   if (db == "ncbi") reads[[1]]$rname <- identify_rnames(reads)
@@ -503,7 +505,6 @@ metascope_id <- function(input_file, input_type = "csv.gz",
     }
   }
 
-
   if (update_bam) {
     combined_distinct <- results[[2]] |>
       dplyr::mutate(qname_names = read_names[.data$qname],
@@ -526,17 +527,18 @@ metascope_id <- function(input_file, input_type = "csv.gz",
     Rsamtools::indexBam(files = input_file)
     input_bam <- Rsamtools::BamFile(input_file, index = input_file,
                                     yieldSize = 100000000)
-    Rsamtools::filterBam(input_bam, destination = bam_out, filter = filter_which)
-    Rsamtools::indexBam(files = bam_out)
-    # Delete old bam file
-    old_bam <- file.path(tmp_dir, paste0(out_base, ".bam"))
-    old_bam_bai <- file.path(tmp_dir, paste0(out_base, ".bam.bai"))
-    if (file.exists(old_bam)) {
-      #Delete file if it exists
-      file.remove(old_bam)
-      file.remove(old_bam_bai)
+    if (length(list(filter_which)) != length(list(bam_out))) {
+      message("update_bam unable to filter. Step skipped")
+    } else {
+      Rsamtools::filterBam(input_bam, destination = list(bam_out), filter = list(filter_which))
+      Rsamtools::indexBam(files = bam_out)
+      message("Updated bam file written to ", bam_out)
+      # Delete old bam file
+      old_bam <- file.path(tmp_dir, paste0(out_base, ".bam"))
+      old_bam_bai <- file.path(tmp_dir, paste0(out_base, ".bam.bai"))
+      if (file.exists(old_bam)) file.remove(old_bam)
+      if (file.exists(old_bam_bai)) file.remove(old_bam_bai)
     }
-    
   }
   # Plotting genome locations
   num_plot <- num_species_plot
@@ -552,5 +554,3 @@ metascope_id <- function(input_file, input_type = "csv.gz",
   } else if (!quiet) message("No coverage plots created")
   return(metascope_id_file)
 }
-
-
